+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - request for user avatar compression
 - notification bell with notification list
 - added text if there are no notifications
+- if you hover the mouse over a floating notification, it will not hide automatically
 
 ## [0.4.0]